requirements.txt
discord.py[voice]
requests
youtube_dl
gtts
peewee
PyMySQL[rsa]